better error codes

+++ src/disk.cpp
@@ -17,25 +17,25 @@
  *  Disk * mem will be a char array of EMULATOR_SIZE items
  */
 Disk * gros_open_disk() {
-    int result;
+    int    result;
     Disk * disk = new Disk();
     disk->size  = EMULATOR_SIZE;
-    disk->fd    = open("grosfs.filesystem", O_RDWR | O_CREAT, (mode_t)0600);
-    if (disk->fd == -1) {
-        printf("Could not open device for file system..\n");
-        exit(1);
+    disk->fd = open( "grosfs.filesystem", O_RDWR | O_CREAT, ( mode_t ) 0600 );
+    if( disk->fd == -1 ) {
+        printf( "Could not open device for file system..\n" );
+        exit( 1 );
     }
-    result = lseek(disk->fd, EMULATOR_SIZE, SEEK_SET);
-    if (result == -1) {
-        close(disk->fd);
-        printf("Could not extend file to desired file system size..\n");
-        exit(1);
+    result = ( int ) lseek( disk->fd, EMULATOR_SIZE, SEEK_SET );
+    if( result == -1 ) {
+        close( disk->fd );
+        printf( "Could not extend file to desired file system size..\n" );
+        exit( 1 );
     }
-    result = write(disk->fd, "", 1);
-    if (result < 0) {
-        close(disk->fd);
-        printf("Could not write a byte to the end of the file..\n");
-        exit(1);
+    result = ( int ) write( disk->fd, "", 1 );
+    if( result < 0 ) {
+        close( disk->fd );
+        printf( "Could not write a byte to the end of the file..\n" );
+        exit( 1 );
     };
     return disk;
 }
@@ -47,7 +47,7 @@ Disk * gros_open_disk() {
  * @param Disk * disk    The pointer to the disk to close
  */
 void gros_close_disk( Disk * disk ) {
-    close(disk->fd);
+    close( disk->fd );
     delete disk;
 }
 
@@ -67,8 +67,8 @@ int gros_read_block( Disk * disk, int block_num, char * buf ) {
     if( ( byte_offset + BLOCK_SIZE ) > disk->size )
         return -1;
 
-    lseek(disk->fd, byte_offset, SEEK_SET);
-    read(disk->fd, buf, BLOCK_SIZE);
+    lseek( disk->fd, byte_offset, SEEK_SET );
+    read( disk->fd, buf, BLOCK_SIZE );
 
     return 0;
 }
@@ -89,8 +89,8 @@ int gros_write_block( Disk * disk, int block_num, char * buf ) {
     if( ( byte_offset + BLOCK_SIZE ) > disk->size )
         return -1;
 
-    lseek(disk->fd, byte_offset, SEEK_SET);
-    write(disk->fd, buf, BLOCK_SIZE);
+    lseek( disk->fd, byte_offset, SEEK_SET );
+    write( disk->fd, buf, BLOCK_SIZE );
 
     return 0;
 }
@@ -103,9 +103,9 @@ TEST_CASE( "Disk emulator can be accessed properly", "[disk]" ) {
 
     REQUIRE( disk->size == EMULATOR_SIZE );
     REQUIRE( disk->fd != -1 );
-    lseek(disk->fd, 0L, SEEK_END);
-    fstat(disk->fd, &stbuf);
-    REQUIRE( stbuf.st_size == disk->size+1 );
+    lseek( disk->fd, 0L, SEEK_END );
+    fstat( disk->fd, &stbuf );
+    REQUIRE( stbuf.st_size == disk->size + 1 );
 
     SECTION( "gros_read_block will gros_read out BLOCK_SIZE bytes" ) {
         int ret = gros_read_block( disk, 0, buf );
@@ -117,7 +117,8 @@ TEST_CASE( "Disk emulator can be accessed properly", "[disk]" ) {
         REQUIRE( ret != 0 );
     }
 
-    SECTION( "gros_read_block will fail on a block number that is out of range" ) {
+    SECTION(
+            "gros_read_block will fail on a block number that is out of range" ) {
         int block_num = ( disk->size / BLOCK_SIZE ) + 1;
         int ret = gros_read_block( disk, block_num, buf );
         REQUIRE( ret != 0 );
@@ -133,12 +134,13 @@ TEST_CASE( "Disk emulator can be accessed properly", "[disk]" ) {
         REQUIRE( ret != 0 );
     }
 
-    SECTION( "gros_write_block will fail on a block number that is out of range" ) {
+    SECTION(
+            "gros_write_block will fail on a block number that is out of range" ) {
         int block_num = ( disk->size / BLOCK_SIZE ) + 1;
         int ret = gros_write_block( disk, block_num, buf );
         REQUIRE( ret != 0 );
     }
-    gros_close_disk(disk);
+    gros_close_disk( disk );
 }
 
 TEST_CASE( "Testing Catch", "[test]" ) {
